refactor: 불필요한 파일 삭제 및 이름 변경
