+++ apps/frontend-menu-viewer/src/hooks/useFetchMenuDetailQuery.ts
@@ -1,4 +1,4 @@
-import { menuAPI } from '@/apis/menu/menu.api';
+import { menuAPI } from '@/services/menu/menu.api';
 import { useQuery } from '@tanstack/react-query';
 
 export function useFetchMenuDetailQuery(menuId: string) {
