+++ apps/frontend-menu-viewer/src/hooks/useFetchStoreInfoQuery.ts
@@ -1,4 +1,4 @@
-import { storeAPI } from '@/apis/store/store.api';
+import { storeAPI } from '@/services/store/store.api';
 import { useQuery } from '@tanstack/react-query';
 
 export default function useFetchStoreInfoQuery(storeId: string) {
