+++ apps/frontend-menu-viewer/src/mocks/api/store.ts
@@ -1,5 +1,5 @@
 import { StoreInfoData } from './data/StoreInfoData';
-import { API_BASE_URL } from '@/apis/api';
+import { API_BASE_URL } from '@/services/api';
 import { delay, http, HttpResponse } from 'msw';
 
 export const storeHandlers = [
