+++ apps/frontend-menu-viewer/src/utilities/formatNumber.ts
