+++ apps/frontend-menu-viewer/src/utilities/parseUserAgent.ts
