--- apps/frontend-user/.dockerignore
@@ -1,10 +1,0 @@
-node_modules
-.git
-dist
-.vscode
-apps/*/node_modules
-apps/*/dist
-apps/*/.next
-pnpm-lock.yaml.backup
-*.log
-Dockerfile
