--- apps/frontend-user/eslint.config.mjs
@@ -1,55 +1,0 @@
-import { FlatCompat } from '@eslint/eslintrc';
-import js from '@eslint/js';
-import importPlugin from 'eslint-plugin-import';
-import globals from 'globals';
-import { dirname } from 'path';
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-const compat = new FlatCompat({
-  baseDirectory: __dirname,
-});
-
-const eslintConfig = [
-  js.configs.recommended,
-  ...compat.extends('next/core-web-vitals', 'next/typescript'),
-  {
-    plugins: {
-      import: importPlugin,
-    },
-    languageOptions: {
-      globals: {
-        ...globals.browser,
-        ...globals.es2021,
-      },
-    },
-    rules: {
-      'react-hooks/exhaustive-deps': 'warn',
-      'react/no-unescaped-entities': 'off',
-      '@next/next/no-img-element': 'off',
-      'import/order': [
-        'error',
-        {
-          groups: [
-            'builtin',
-            'external',
-            'internal',
-            'parent',
-            'sibling',
-            'index',
-          ],
-          'newlines-between': 'always',
-          alphabetize: {
-            order: 'asc',
-            caseInsensitive: true,
-          },
-        },
-      ],
-      'import/no-unresolved': 'off',
-    },
-  },
-];
-
-export default eslintConfig;
