--- apps/frontend-user/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "frontend-user",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev --turbopack -p 3001",
-    "build": "next build",
-    "start": "next start -p 3001 -H 0.0.0.0",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@tanstack/react-query": "^5.81.5",
-    "axios": "^1.10.0",
-    "next": "15.3.4",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-toastify": "^11.0.5",
-    "tokens": "workspace:*",
-    "zustand": "^5.0.6"
-  },
-  "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@eslint/js": "^9.29.0",
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "eslint": "^9",
-    "eslint-config-next": "15.3.4",
-    "eslint-plugin-import": "^2.32.0",
-    "globals": "^16.2.0",
-    "tailwindcss": "^4",
-    "typescript": "^5"
-  }
-}
