--- apps/frontend-user/src/styles/global.css
@@ -1,3 +1,0 @@
-@import 'tailwindcss';
-@import 'tokens/src/colors.css';
-@import 'tokens/src/typography.css';
